+++ src/main/java/com/ooa1769/bs/BSApp.java
@@ -2,10 +2,12 @@ package com.ooa1769.bs;
 
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.context.annotation.EnableAspectJAutoProxy;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @SpringBootApplication
 @EnableJpaAuditing
+@EnableAspectJAutoProxy(proxyTargetClass = true)
 public class BSApp {
 
     public static void main(String[] args) {
