--- src/main/java/com/ooa1769/bs/member/repository/MemberRepository.java
@@ -1,11 +1,0 @@
-package com.ooa1769.bs.member.repository;
-
-import com.ooa1769.bs.member.domain.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-
-    Optional<Member> findByEmail(String email);
-}
