+++ src/main/java/com/ooa1769/bs/respository/member/MemberRepository.java
@@ -1,0 +1,11 @@
+package com.ooa1769.bs.respository.member;
+
+import com.ooa1769.bs.domain.member.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+
+    Optional<Member> findByEmail(String email);
+}
