+++ src/main/java/com/ooa1769/bs/support/security/LoginMember.java
@@ -1,0 +1,11 @@
+package com.ooa1769.bs.support.security;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface LoginMember {
+}
