+++ src/main/java/com/ooa1769/bs/support/security/UnAuthenticationException.java
@@ -1,0 +1,26 @@
+package com.ooa1769.bs.support.security;
+
+public class UnAuthenticationException extends Exception {
+    private static final long serialVersionUID = 1L;
+
+    public UnAuthenticationException() {
+        super();
+    }
+
+    public UnAuthenticationException(String message, Throwable cause, boolean enableSuppression,
+                                     boolean writableStackTrace) {
+        super(message, cause, enableSuppression, writableStackTrace);
+    }
+
+    public UnAuthenticationException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public UnAuthenticationException(String message) {
+        super(message);
+    }
+
+    public UnAuthenticationException(Throwable cause) {
+        super(cause);
+    }
+}
