+++ src/main/resources/templates/loginPage.html
@@ -28,7 +28,6 @@
             <input type="submit" class="btn btn-primary" value="Sign In" />
             <a class="btn btn-default" th:href="@{/signup}">Sign Up</a>
             <br/>
-            <a th:href="@{/forgotPassword}">Reset Password</a>
         </div>
     </form>
 </div>
