+++ src/test/java/com/ooa1769/bs/domain/member/MemberTest.java
@@ -1,0 +1,5 @@
+package com.ooa1769.bs.domain.member;
+
+public class MemberTest {
+
+}
