--- src/test/java/com/ooa1769/bs/member/domain/MemberTest.java
@@ -1,5 +1,0 @@
-package com.ooa1769.bs.member.domain;
-
-public class MemberTest {
-
-}
