+++ src/test/java/com/ooa1769/bs/service/member/MemberServiceTest.java
@@ -1,0 +1,11 @@
+package com.ooa1769.bs.service.member;
+
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class MemberServiceTest {
+
+}
